Update snmp_entity_table.py
At vodafone it was seen that a device was not able to respond to snmpget but snmpwalks over the same range were able to get the value. Their Huawei contact that looked at the issue insisted there was no issue and it was somehow user error of some sort. To avoid this, I instead got the tables and then referenced them in the function instead of individually snmpgetting each value.