Update huawei_save_flow.py
The devices were having an issue trying to restore the configurations. The issue was caused because we were saving the configuration to "quali_run_config.cfg" then copying it over to a new file. This was causing some reference to not be created internally on the device to allow the file to be loaded as a valid configuration.